+++ R/0D/libs/sdw/CanvasPoint.h
@@ -8,14 +8,14 @@ class CanvasPoint
   public:
     float x;
     float y;
-    double depth;
+    float depth;
     float brightness;
     TexturePoint texturePoint;
 
     CanvasPoint();
     CanvasPoint(float xPos, float yPos);
 
-    CanvasPoint(float xPos, float yPos, double pointDepth);
+    CanvasPoint(float xPos, float yPos, float pointDepth);
 
     //Undefined behaviour to add or subtract -1 point depth
     CanvasPoint operator-(const CanvasPoint& c);
